+++ src/it/java/org/seedstack/validation/internal/ValidationServiceIT.java
@@ -11,16 +11,13 @@ import org.assertj.core.api.Assertions;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.seedstack.seed.it.SeedITRunner;
-import org.seedstack.validation.ValidationException;
-import org.seedstack.validation.ValidationService;
 import org.seedstack.validation.internal.pojo.Bean;
 import org.seedstack.validation.internal.pojo.MyImpl;
 import org.seedstack.validation.internal.pojo.Pojo;
 import org.seedstack.validation.internal.pojo.PojoWithDeepValidation;
 
 import javax.inject.Inject;
-import javax.validation.ConstraintViolation;
-import java.util.Set;
+import javax.validation.ConstraintViolationException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -39,16 +36,10 @@ public class ValidationServiceIT {
     public void throwsExceptionOnInvalidPojo() {
         try {
             validationService.staticallyHandle(new Pojo(Pojo.State.INVALID));
-            Assertions.failBecauseExceptionWasNotThrown(ValidationException.class);
-        } catch (ValidationException validationException) {
-            expectViolations(validationException, 3);
+            Assertions.failBecauseExceptionWasNotThrown(ConstraintViolationException.class);
+        } catch (ConstraintViolationException exception) {
+            assertThat(exception.getConstraintViolations()).hasSize(3);
         }
-    }
-
-    private void expectViolations(ValidationException validationException, int expected) {
-        Set<ConstraintViolation<?>> constraintViolations = validationException
-                .get(ValidationService.JAVAX_VALIDATION_CONSTRAINT_VIOLATIONS);
-        assertThat(constraintViolations).hasSize(expected);
     }
 
     @Test
@@ -60,9 +51,9 @@ public class ValidationServiceIT {
     public void throwsExceptionOnDeepValidation() {
         try {
             validationService.staticallyHandle(new PojoWithDeepValidation());
-            Assertions.failBecauseExceptionWasNotThrown(ValidationException.class);
-        } catch (ValidationException validationException) {
-            expectViolations(validationException, 4);
+            Assertions.failBecauseExceptionWasNotThrown(ConstraintViolationException.class);
+        } catch (ConstraintViolationException exception) {
+            assertThat(exception.getConstraintViolations()).hasSize(4);
         }
     }
 
@@ -70,9 +61,9 @@ public class ValidationServiceIT {
     public void validationShouldWorkOnInterface() {
         try {
             validationService.staticallyHandle(new MyImpl());
-            Assertions.failBecauseExceptionWasNotThrown(ValidationException.class);
-        } catch (ValidationException validationException) {
-            expectViolations(validationException, 1);
+            Assertions.failBecauseExceptionWasNotThrown(ConstraintViolationException.class);
+        } catch (ConstraintViolationException exception) {
+            assertThat(exception.getConstraintViolations()).hasSize(1);
         }
     }
 
@@ -82,9 +73,9 @@ public class ValidationServiceIT {
             Bean candidate = new Bean();
             candidate.setHour(25);
             validationService.staticallyHandle(candidate);
-            Assertions.failBecauseExceptionWasNotThrown(ValidationException.class);
-        } catch (ValidationException validationException) {
-            expectViolations(validationException, 1);
+            Assertions.failBecauseExceptionWasNotThrown(ConstraintViolationException.class);
+        } catch (ConstraintViolationException exception) {
+            assertThat(exception.getConstraintViolations()).hasSize(1);
         }
     }
 }