+++ src/main/java/org/seedstack/validation/internal/ValidationExceptionMapper.java
@@ -7,17 +7,16 @@
  */
 package org.seedstack.validation.internal;
 
-import org.seedstack.validation.ValidationException;
-
+import javax.validation.ConstraintViolationException;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
 import javax.ws.rs.ext.Provider;
 
 @Provider
-public class ValidationExceptionMapper implements ExceptionMapper<ValidationException> {
+public class ValidationExceptionMapper implements ExceptionMapper<ConstraintViolationException> {
 
     @Override
-    public Response toResponse(ValidationException exception) {
+    public Response toResponse(ConstraintViolationException exception) {
         return Response.status(Response.Status.BAD_REQUEST).build();
     }
 }