+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Version 2.2.0 (2016-01-28)
 
-* [new] An ExceptionMapper for ValidationException is provided by default.
+* [new] An `ExceptionMapper` for `ConstraintViolationException` is provided by default.
+* [new] The `ValidatorFactory` is now injectable.
+* [brk] The `ValidationService` is deprecated.
+* [brk] The `org.seedstack.validation.ValidationException` is replaced by the standard `ConstraintViolationException`.